Function moved to suggested routes controller as suggested in the review
